+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Imp/EpochSpec.hs
@@ -421,7 +421,7 @@ treasuryWithdrawalExpectation extraWithdrawals = do
   donateToTreasury withdrawalAmount
   committeeHotCreds <- registerInitialCommittee
   (dRepCred, _, _) <- setupSingleDRep 1_000_000
-  treasuryStart <- getsNES $ nesEsL . esAccountStateL . asTreasuryL
+  treasuryStart <- getsNES $ nesEsL . esChainAccountStateL . asTreasuryL
   treasuryStart `shouldBe` withdrawalAmount
   rewardAccount <- registerRewardAccount
   govPolicy <- getGovPolicy
@@ -436,7 +436,7 @@ treasuryWithdrawalExpectation extraWithdrawals = do
     getReward (raCredential rewardAccount) `shouldReturn` mempty
   passEpoch -- 2nd epoch crossing enacts all the ratified actions
   expectMissingGovActionId govActionId
-  treasuryEnd <- getsNES $ nesEsL . esAccountStateL . asTreasuryL
+  treasuryEnd <- getsNES $ nesEsL . esChainAccountStateL . asTreasuryL
   impAnn "Withdrawal deducted from treasury" $
     treasuryStart <-> treasuryEnd `shouldBe` withdrawalAmount
   impAnn "Withdrawal received by reward account" $
@@ -445,7 +445,7 @@ treasuryWithdrawalExpectation extraWithdrawals = do
 depositMovesToTreasuryWhenStakingAddressUnregisters :: ConwayEraImp era => ImpTestM era ()
 depositMovesToTreasuryWhenStakingAddressUnregisters = do
   disableTreasuryExpansion
-  initialTreasury <- getsNES $ nesEsL . esAccountStateL . asTreasuryL
+  initialTreasury <- getsNES $ nesEsL . esChainAccountStateL . asTreasuryL
   modifyPParams $ \pp ->
     pp
       & ppGovActionLifetimeL .~ EpochInterval 8