+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Rules/Mir.hs
@@ -27,7 +27,7 @@ import Cardano.Ledger.Keys (KeyRole (..))
 import Cardano.Ledger.Shelley.Era (ShelleyMIR)
 import Cardano.Ledger.Shelley.Governance (EraGov)
 import Cardano.Ledger.Shelley.LedgerState (
-  AccountState (..),
+  ChainAccountState (..),
   EpochState,
   InstantaneousRewards (..),
   certDStateL,
@@ -35,7 +35,7 @@ import Cardano.Ledger.Shelley.LedgerState (
   dsIRewards,
   dsIRewardsL,
   dsUnifiedL,
-  esAccountState,
+  esChainAccountState,
   esLState,
   esLStateL,
   esNonMyopic,
@@ -118,7 +118,7 @@ mirTransition = do
   TRC
     ( _
       , es@EpochState
-          { esAccountState = acnt
+          { esChainAccountState = chainAccountState
           , esSnapshots = ss
           , esLState = ls
           , esNonMyopic = nm
@@ -131,8 +131,8 @@ mirTransition = do
       dpState = ls ^. lsCertStateL
       ds = dpState ^. certDStateL
       rewards' = rewards ds
-      reserves = asReserves acnt
-      treasury = asTreasury acnt
+      reserves = casReserves chainAccountState
+      treasury = casTreasury chainAccountState
       irwdR = rewards' UM.◁ iRReserves (dsIRewards ds) :: Map.Map (Credential 'Staking) Coin
       irwdT = rewards' UM.◁ iRTreasury (dsIRewards ds) :: Map.Map (Credential 'Staking) Coin
       totR = fold irwdR
@@ -146,9 +146,9 @@ mirTransition = do
       tellEvent $ MirTransfer ((dsIRewards ds) {iRReserves = irwdR, iRTreasury = irwdT})
       pure $
         EpochState
-          acnt
-            { asReserves = availableReserves <-> totR
-            , asTreasury = availableTreasury <-> totT
+          ChainAccountState
+            { casReserves = availableReserves <-> totR
+            , casTreasury = availableTreasury <-> totT
             }
           ( ls
               & lsCertStateL . certDStateL . dsUnifiedL .~ (rewards' UM.∪+ Map.map compactCoinOrError update)
@@ -166,7 +166,7 @@ mirTransition = do
           availableTreasury
       pure $
         EpochState
-          acnt
+          chainAccountState
           ( ls
               & lsCertStateL . certDStateL . dsIRewardsL .~ emptyInstantaneousRewards
           )