+++ libs/cardano-ledger-core/testlib/Test/Cardano/Ledger/TreeDiff.hs
@@ -178,7 +178,7 @@ instance ToExpr TxIn
 
 instance ToExpr TxId
 
-instance ToExpr AccountState
+instance ToExpr ChainAccountState
 
 -- CertState
 instance ToExpr DRep