+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Generic/PrettyCore.hs
@@ -188,7 +188,7 @@ import Cardano.Ledger.Shelley.AdaPots (
  )
 import Cardano.Ledger.Shelley.Core
 import Cardano.Ledger.Shelley.LedgerState (
-  AccountState (..),
+  ChainAccountState (..),
   DState (..),
   EpochState (..),
   EraCertState (..),
@@ -3291,10 +3291,10 @@ showProtver :: ProtVer -> String
 showProtver (ProtVer x y) = "(" ++ show x ++ " " ++ show y ++ ")"
 
 pcEpochState :: Reflect era => Proof era -> EpochState era -> PDoc
-pcEpochState proof es@(EpochState (AccountState tre res) ls sss nonmy) =
+pcEpochState proof es@(EpochState (ChainAccountState tre res) ls sss nonmy) =
   ppRecord
     "EpochState"
-    [ ("AccountState", ppRecord' "" [("treasury", pcCoin tre), ("reserves", pcCoin res)])
+    [ ("ChainAccountState", ppRecord' "" [("treasury", pcCoin tre), ("reserves", pcCoin res)])
     , ("LedgerState", pcLedgerState proof ls)
     , ("SnapShots", pcSnapShots sss)
     , ("NonMyopic", ppNonMyopic nonmy)
@@ -3304,18 +3304,18 @@ pcEpochState proof es@(EpochState (AccountState tre res) ls sss nonmy) =
 instance Reflect era => PrettyA (EpochState era) where
   prettyA = pcEpochState reify
 
-pcAccountState :: AccountState -> PDoc
-pcAccountState (AccountState tr re) = ppRecord' "" [("treasury", pcCoin tr), ("reserves", pcCoin re)]
+pcChainAccountState :: ChainAccountState -> PDoc
+pcChainAccountState (ChainAccountState tr re) = ppRecord' "" [("treasury", pcCoin tr), ("reserves", pcCoin re)]
 
-instance PrettyA AccountState where
-  prettyA = pcAccountState
+instance PrettyA ChainAccountState where
+  prettyA = pcChainAccountState
 
 -- | Like pcEpochState.but it only prints a summary of the UTxO
 psEpochState :: Reflect era => Proof era -> EpochState era -> PDoc
-psEpochState proof es@(EpochState (AccountState tre res) ls sss _) =
+psEpochState proof es@(EpochState (ChainAccountState tre res) ls sss _) =
   ppRecord
     "EpochState"
-    [ ("AccountState", ppRecord' "" [("treasury", pcCoin tre), ("reserves", pcCoin res)])
+    [ ("ChainAccountState", ppRecord' "" [("treasury", pcCoin tre), ("reserves", pcCoin res)])
     , ("LedgerState", psLedgerState proof ls)
     , ("SnapShots", pcSnapShots sss)
     , ("AdaPots", pcAdaPot es)